% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funclib_plot.R
\name{batchCompare}
\alias{batchCompare}
\title{batchCompare}
\usage{
batchCompare(datalist,"upset","read",mycolors=c("#b681bd","#377eb8"))
}
\arguments{
\item{datalist}{Multiple sets of analysis results stored in list format.}

\item{format}{Plot format, "upset" or "veen".}

\item{dimension}{The dimension to compare, "gene", "tail", "PACds" or "read".}

\item{mycolors}{Drawing color card, must be specified, no default value. The
color number is the same as the number of data sets.}

\item{rep}{Logical variable. When plotting the comparison of poly(A) tail
 length distribution for different batches of data, the parameter 'rep' can
 be used to specify whether to display data from different biological
 replicates separately. If the parameter 'rep' is true, the data of
 different repetitions are displayed separately, otherwise they are not
 separated.
@param findOvpPACds Logical variable. Indicates whether to output
  overlapping PACds, defaults to 'False'.
@param annotateByKnownPAC Logical variable. annotateByKnownPAC returns
  overlapping status and min distance with given PACdatasets.}
}
\value{
Return the comparison result graph.
}
\description{
The batchCompare function can be used to compare the calculation
  results of multiple different batches of data.
}
\details{
The batchCompare function provides the ability to quickly compare
  and visualize tail quantification results from multiple different batches
  of data. Users can use this function to compare computational results from
  multiple different datasets, such as one from NGS sequencing technology and
  another from PACBIO sequencing technology. batchCompare provides three
  comparable dimensions, namely "gene", "tail", "PACds" or "read".
}
