% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funclib_PolyAtailor.R
\name{ccf}
\alias{ccf}
\title{ccf}
\usage{
ccf(ref, PAdf, fromFormat, toFormat)
}
\arguments{
\item{ref}{Reference table of chromosome information from NCBI database.}

\item{PAdf}{Files that need to convert chromosome names, dataframe format.}

\item{fromFormat}{Primitive chromosome nomenclature.}

\item{toFormat}{Target format of chromosome name.}
}
\value{
This function returns a data frame with the chromosome name changed.
}
\description{
Convert chromosome name format.
}
\details{
For various reasons, we sometimes need to convert the format of
  chromosome names in files. This function can help users convert the format.
}
\examples{
ref <- read.table("./data/HELA1/GRCH38p13_sequences_report.txt",
  header = T, sep = "\t", blank.lines.skip = T
)
PAdf <- read.table("***")
result <- ccf(ref, PAdf, "RefSeq.Accn", "UCSC.style.name")
}
