% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funclib_polyAtailor.R
\name{diffPALdpPA}
\alias{diffPALdpPA}
\title{diffPALdpPA}
\usage{
diffPALdpPA(PAdf, PALdf, gff, d)
}
\arguments{
\item{PAdf}{A dataframe of PA infomation.}

\item{PALdf}{A dataframe that contains the length information of all reads
tails.}

\item{gff}{Genome annotation stored in a GFF/GTF file or a TXDB R object can
be used for annotating PACs. Please refer to movAPA for details.}

\item{d}{distance to group nearby PACds, default is 24 nt.}
}
\value{
The function returns a dataframe of result. The result data frame
  contains 7 column contents, the first column is geneID, and the second and
  third columns are the median tail lengths under different conditions. The
  remaining four columns are p-values calculated by different difference
  significance test methods, each column corresponds to a difference
  significance test method, in which "lose" or "NAN" means that for some
  reason (probably too little data) this method has not been able to
  calculate the exact p-values.
}
\description{
Analyzing whether there is a significant difference
  in polyA tail length between the proximal and distal PA sites of genes with
  two PACs.
}
\details{
This function analyzes whether there is a significant difference
  in polyA tail length between the proximal and distal PA sites of genes with
  two PACs.
}
