% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funclib_plot.R
\name{DSA_UpsetPlot}
\alias{DSA_UpsetPlot}
\title{DSA_UpsetPlot}
\usage{
DSA_UpsetPlot(DSA_result)
}
\arguments{
\item{DSA_result}{The output of PALdsa().}
}
\value{
A UpsetPlot.
}
\description{
Draw a Upset diagram of the result of differential tail length
  analysis.
}
\details{
Overlapping analysis of the results of differential tail length
  analysis of the four difference significance testing methods can help users
  quickly determine which results of differential tail length analysis are
  highly confident. The consensus results of the four methods are marked in
  red to represent highly confident tail length genes.
}
\seealso{
[DSA_ViolinPlot()] to draw a violin diagram of the result of
  differential tail length analysis.

Other Visualization functions: 
\code{\link{DSA_ViolinPlot}()},
\code{\link{PALdsa}()},
\code{\link{nonAanalysis}()},
\code{\link{plotGenePAnumbers}()},
\code{\link{plotPADistribution}()},
\code{\link{plotPALDistribution}()},
\code{\link{plotPASignals}()},
\code{\link{tailViso}()}
}
\concept{Visualization functions}
