% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funclib_plot.R
\name{DSA_ViolinPlot}
\alias{DSA_ViolinPlot}
\title{DSA_ViolinPlot}
\usage{
DSA_ViolinPlot(DSA_result, SAoDMethod)
}
\arguments{
\item{DSA_result}{The output of PALdsa().}

\item{SAoDMethod}{one of "KS", "MWU", "ME", "Wilcox" and "ALL".}
}
\value{
A ViolinPlot.
}
\description{
Draw a violin diagram of the result of differential tail length
  analysis.
}
\details{
The tail length difference analysis results of the four difference
  significance test methods are displayed through the violin diagram. The
  tail length distribution under different conditions is represented by
  different colors to facilitate users to view the differences.
}
\seealso{
[DSA_UpsetPlot()] to draw a Upset diagram of the result of
  differential tail length analysis.

Other Visualization functions: 
\code{\link{DSA_UpsetPlot}()},
\code{\link{PALdsa}()},
\code{\link{nonAanalysis}()},
\code{\link{plotGenePAnumbers}()},
\code{\link{plotPADistribution}()},
\code{\link{plotPALDistribution}()},
\code{\link{plotPASignals}()},
\code{\link{tailViso}()}
}
\concept{Visualization functions}
