% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funclib_polyAtailor.R
\name{findAndAnnoPAs}
\alias{findAndAnnoPAs}
\title{findAndAnnoPAs}
\usage{
findAndAnnoPAs(
  chrinfo,
  bamfile,
  resultpath,
  bsgenome,
  gffFile,
  sample,
  mergePAs,
  d
)
}
\arguments{
\item{chrinfo}{The path of chrinfo file.}

\item{bamfile}{The path of bam file.}

\item{resultpath}{The path of result file.}

\item{bsgenome}{BSgenome package for your species, please refer to movAPA
package for details.}

\item{gffFile}{Genome annotation stored in a GFF/GTF file or a TXDB R object
can be used for annotating PACs. Please refer to movAPA for details.}

\item{sample}{The sample name.}

\item{mergePAs}{TRUE/FALSE. If TRUE, then will mergePACds groups nearby PACs
from single/multiple PACdataset objects.}

\item{d}{distance to group nearby PACds, default is 24 nt.}
}
\value{
The function returns data in PACds format with the annotation
  information for all PA sites. In addition, we will also generate a
  "pac_data.txt" file to store the PA annotation table under the result path
  you specified, and a "pac_data.coldata.txt" file to store the additional
  comment information.And an "ACTGpdf.PDF" file, which holds a legend of the
  base composition around the PACs, has two images representing the base
  composition of the plus and minus chains.
}
\description{
Find and annotate PA sites.
}
\examples{
library(movAPA)
# Deciphering gff files
library(TxDb.Mmusculus.UCSC.mm10.knownGene)
# Deciphering genome files
library("BSgenome.Mmusculus.UCSC.mm10")
bsgenome = BSgenome.Mmusculus.UCSC.mm10
bamfilepath = system.file("extdata", "./GV_algin/PAIso-GV1.sorted.bam",
package = "PolyAtailor", mustWork = TRUE)
chrinfopath = system.file("extdata", "./GV_algin/chrinfo.txt", package =
"PolyAtailor", mustWork = TRUE)
resultpath = "./"
# Annotated PA site
PAs <- findAndAnnoPAs(bamfile=bamfilepath,chrinfo=chrinfopath,resultpath=resultpath,bsgenome=bsgenome,gffFile = TxDb.Mmusculus.UCSC.mm10.knownGene,sample="GV1",mergePAs=T,d=24)
}
\seealso{
Other poly(A) sites detection functions: 
\code{\link{findPAs}()}
}
\concept{poly(A) sites detection functions}
