% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funclib_polyAtailor.R
\name{findChrTails}
\alias{findChrTails}
\title{findChrTails}
\usage{
findChrTails(bamfile, which, what, count)
}
\arguments{
\item{bamfile}{The path of bam file.}

\item{which}{Refer to ScanBamParam parameters for Rsamtools for explanation.}

\item{what}{Refer to ScanBamParam parameters for Rsamtools for explanation.}

\item{count}{Boolean parameter. If count=T, the PA list after the count will
be returned. There will be one more column of count in the PA list. If
count=F, all PA site information is returned without counting.}
}
\value{
Returns a PA list annotated with chromosome information, chain
  orientation, and reference genomic coordinates.
}
\description{
Each chromosome was traversed and PA sites were identified by
  a-rich at the end. Note that the BAM file needs to be indexed.
}
