% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infomerge.R
\name{infomerge}
\alias{infomerge}
\title{infomerge}
\usage{
infomerge(tailinfo, PAinfo)
}
\arguments{
\item{tailinfo}{A dataframe that contains the tail information of all reads.}

\item{PAinfo}{A dataframe that contains the PA information of all reads.}
}
\value{
This function returns a dataframe that contains both PA information
  (including PA gene, type based, chromosome, and so on) and tail information.
}
\description{
Merging PA information and tail information.
}
\details{
Although we have identified the PA site and tail information for
  each sequence, the two tables are currently independent. We need to use
  the common column (reads_num) of the two tables to merge the data to
  produce the final data table.
}
\examples{
tailinfo = "the/path/of/tailinfo.txt"
PAinfo = "the/path/of/PAinfo.txt"
finaldata <- infomerge(tailinfo,PAinfo)
}
