% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funclib_plot.R
\name{nonAanalysis}
\alias{nonAanalysis}
\title{nonAanalysis}
\usage{
nonAanalysis(tailinfo, justOneTail)
}
\arguments{
\item{tailinfo}{A tail information table containing at least PAL, sample
name, tail sequence, chain direction, and gene infomation.}

\item{justOneTail}{Boolean value. Indicates whether to analyze only data of
type one-tail, which defaults to T.}
}
\value{
A large list containing quantitative data tables and statistical
  graphs.The two figures respectively represent the column chart of the
  number and proportion of reads with non-A bases in each sample and the
  specific occurrence frequency of non-A bases in each sample.
}
\description{
Quantify and Statistics the nonA base in tail.
}
\details{
After users input the tail information table, we will quantify the
  non-A base content in each tail, and count the number of reads with non-A
  base tail in each sample and the proportion of non-A base content in each
  sample.
}
\examples{
data(AnnotedTails)
re <- nonAanalysis(AnnotedTails)
re$p1
re$p2
re$p3
nonAre <- re$nonAinfo
}
\seealso{
[tailViso()] to visualize tails using heat maps and logos.

Other Visualization functions: 
\code{\link{DSA_UpsetPlot}()},
\code{\link{DSA_ViolinPlot}()},
\code{\link{PALdsa}()},
\code{\link{plotGenePAnumbers}()},
\code{\link{plotPADistribution}()},
\code{\link{plotPALDistribution}()},
\code{\link{plotPASignals}()},
\code{\link{tailViso}()}
}
\concept{Visualization functions}
