% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funclib_plot.R
\name{PALdsa}
\alias{PALdsa}
\title{PALdsa}
\usage{
PALdsa(PAdf, PALdf, gff, d, mode, withViolinPlot, withUpsetPlot, SAoDMethod)
}
\arguments{
\item{PAdf}{A dataframe of PA infomation without merge.}

\item{PALdf}{A dataframe that contains the length information of all reads
tails.}

\item{gff}{Genome annotation stored in a GFF/GTF file or a TXDB R object can
be used for annotating PACs. Please refer to movAPA for details.}

\item{d}{distance to group nearby PACds, default is 24 nt.}

\item{mode}{The PAL comparison mode, "PD" refers to comparing the tail
length difference between the proximal and distal PA sites of each gene
with two PACs."2PA" represents a direct comparison of tail length
differences between the two PA loci of genes with two PACs, default for
"PD".}

\item{withViolinPlot}{Logical value variable, Whether to draw a violin
diagram, see function DSA_ViolinPlot() for details, default is true.}

\item{withUpsetPlot}{Logical value variable, Whether to draw a upset
diagram, see function DSA_UpsetPlot() for details, default is true.}

\item{SAoDMethod}{one of "KS", "MWU", "ME", "Wilcox" and "ALL".}
}
\value{
By default, this function returns a list of two graphs and a
  dataframe, or only a dataframe if no drawing operation has been performed.
  The result dataframe contains 7 column contents, the first column is
  geneID, and the second and third columns are the median tail lengths under
  different conditions. The remaining four columns are p-values calculated by
  different difference significance test methods, each column corresponds to
  a difference significance test method, in which "lose" or "NAN" means that
  for some reason (probably too little data) this method has not been able to
  calculate the exact p-values.
}
\description{
Analysis of significant difference of tail length under
  different conditions.
}
\details{
The poly(A) tail length of genes with two PACs was analyzed for
  significant difference.It is feasible to directly analyze whether there is
  a significant difference in tail length between the two PACs, or to
  annotate PACs as the proximal PA site and the distal PA site before
  conducting the significance analysis of tail length difference.
}
\examples{
library(movAPA)
library(TxDb.Mmusculus.UCSC.mm10.knownGene)
gff <- parseGenomeAnnotation(TxDb.Mmusculus.UCSC.mm10.knownGene)
data(AnnotedTails)
files = system.file("extdata", "./output/PAs/PAs.txt", package = "PolyAtailor", mustWork = TRUE)
PAs <- read.table(files,header=TRUE,sep=" ")
diffPAL2PAgenes <-
PALdsa(PAs,AnnotedTails,gff,mode="PD",SAoDMethod="ME",withViolinPlot=TRUE,withUpsetPlot=F)
}
\seealso{
[DSA_ViolinPlot()] to draw a violin diagram of the result of
  differential tail length analysis.

Other Visualization functions: 
\code{\link{DSA_UpsetPlot}()},
\code{\link{DSA_ViolinPlot}()},
\code{\link{nonAanalysis}()},
\code{\link{plotGenePAnumbers}()},
\code{\link{plotPADistribution}()},
\code{\link{plotPALDistribution}()},
\code{\link{plotPASignals}()},
\code{\link{tailViso}()}
}
\concept{Visualization functions}
