% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funclib_plot.R
\name{plotGenePAnumbers}
\alias{plotGenePAnumbers}
\title{plotGenePAnumbers}
\usage{
plotGenePAnumbers(PAdf, resultpath, colorOfBar)
}
\arguments{
\item{PAdf}{A dataframe that contains the PA information of all reads,
Contains at least three columns: PAID, gene, ftr.}

\item{resultpath}{The specified path to save barplot.}

\item{colorOfBar}{The color of barplot, as input in hexadecimal form,
defaults to "#196BE1".}
}
\value{
A barplot.
}
\description{
Plot the gene frequency distribution with different number of PA sites.
}
\details{
Plot the gene frequency distribution with different number of PA
  sites. Genes were divided into five groups, each containing the number of
  PA sites 1, 2, 3, 4, 5 and greater than 5, respectively.
}
\examples{
data(PAs)
p <- plotGenePAnumbers(PAs,"./","#DF7C7D")
}
\seealso{
[plotGenePAnumbers()] to plot the gene frequency distribution with
  different number of PA sites.

Other Visualization functions: 
\code{\link{DSA_UpsetPlot}()},
\code{\link{DSA_ViolinPlot}()},
\code{\link{PALdsa}()},
\code{\link{nonAanalysis}()},
\code{\link{plotPADistribution}()},
\code{\link{plotPALDistribution}()},
\code{\link{plotPASignals}()},
\code{\link{tailViso}()}
}
\concept{Visualization functions}
