% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funclib_plot.R
\name{plotPADistribution}
\alias{plotPADistribution}
\title{plotPADistribution}
\usage{
plotPADistribution(PAdf, resultpath, colorOfBar)
}
\arguments{
\item{PAdf}{A dataframe that contains the PA information of all reads,
Contains at least three columns: PAID, gene, ftr.}

\item{resultpath}{The specified path to save barplot.}

\item{colorOfBar}{The color of barplot, as input in hexadecimal form,
defaults to "#196BE1".}
}
\value{
A barplot.
}
\description{
Plot the gene region distribution of PA sites.
}
\details{
Plot the gene region distribution of PA sites, including "3UTR",
  "5UTR", "CDS", "exon", "intergenic", "intron". Finally, we will save the PA
  locus distribution barpllot in PPTX format to the path you specified.
}
\examples{
data(PAs)
p <- plotPADistribution(PAs,"./","#9BBFDC")
}
\seealso{
[plotGenePAnumbers()] to plot the gene frequency distribution with
  different number of PA sites.
  @export

Other Visualization functions: 
\code{\link{DSA_UpsetPlot}()},
\code{\link{DSA_ViolinPlot}()},
\code{\link{PALdsa}()},
\code{\link{nonAanalysis}()},
\code{\link{plotGenePAnumbers}()},
\code{\link{plotPALDistribution}()},
\code{\link{plotPASignals}()},
\code{\link{tailViso}()}
}
\concept{Visualization functions}
