% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyAtailor.R
\docType{package}
\name{PolyAtailor}
\alias{PolyAtailor}
\title{PolyAtailor}
\description{
This package provides functions for identifying and quantifying
  poly(A) tails of full-length sequencing data from BAM files as well as
  difference significance detection.
}
\details{
This package contains most of the functions required in poly(A) tail
  research, including tail recognition from sequence, tail length
  calculation, non-A base recognition and quantifying tail, PA site
  determination, tail length difference significance detection, basic data
  statistics, and so on.
}
\references{
Ye W, Liu T, Fu H, et al. movAPA: Modeling and visualization of dynamics of alternative polyadenylation across biological samples[J]. Bioinformatics, 2021, 37(16): 2470-2472.   
Liu Y, Nie H, Liu H, et al. Poly (A) inclusive RNA isoform sequencing (PAIso− seq) reveals wide-spread non-adenosine residues within RNA poly (A) tails[J]. Nature communications, 2019, 10(1): 1-13.   
Legnini I, Alles J, Karaiskos N, et al. FLAM-seq: full-length mRNA sequencing reveals principles of poly (A) tail length control[J]. Nature methods, 2019, 16(9): 879-886.
}
