% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funclib_polyAtailor.R
\name{readChrInfo}
\alias{readChrInfo}
\title{Read the ChrInfo of your data.}
\usage{
readChrInfo(file)
}
\arguments{
\item{file}{The path of TXT text of the HEAD information of the Sam file.}
}
\value{
The S4 format data of all the chromosome information.
}
\description{
\code{readChrInfo} returns the S4 format data of all the chromosome information you input.
}
\details{
You need to input the TXT text of the HEAD information of the Sam file after
the alignment, and then the function will calculate and return the name of
the chromosome, the range on the reference genome, the direction of the
sequence, etc
}
