% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funclib_plot.R
\name{tailViso}
\alias{tailViso}
\title{tailViso}
\usage{
tailViso(taildf, tailLen, Ntail, custom, strand, faPath, showLogo, showReadNum)
}
\arguments{
\item{tailLen}{Maximum tail length to draw.}

\item{Ntail}{Number of tails to draw.}

\item{custom}{A dataframe. Custom base color matching information, one column
is the base name, the other column is the corresponding color scheme.}

\item{strand}{The direction of the tail chain, + represents the polyA tail
and - represents the polyT tail.}

\item{faPath}{Tail sequence file path.}

\item{showLogo}{Logical value indicating whether to draw the LOGO diagram,
The default is T.}

\item{showReadNum}{Logical value indicating whether to display the sequence
identifier, default to T.}

\item{tails}{A dataframe. The set of tails that the user wants to visualize
contains at least one set of tail sequences and a unique identifier for
each tail.}
}
\value{
Return the picture.
}
\description{
Visualize tails using heat maps and logos.
}
\details{
Visualize the tail input by the user so that the base composition of
  different tails can be compared intuitively.
}
\examples{
data(taildf)
my_cutstom <- data.frame(names=c("A","C","T","G"),color=c("#3171A5","#4EAA4C","#C9C4C2","#D73D3D"))
p <- tailViso(taildf,tailLen=100,Ntail=20,custom=my_cutstom,strand="-",faPath="D:/",showLogo=T,showReadNum= F)
}
\seealso{
[DSA_ViolinPlot()] to draw a violin diagram of the result of
  differential tail length analysis.

Other Visualization functions: 
\code{\link{DSA_UpsetPlot}()},
\code{\link{DSA_ViolinPlot}()},
\code{\link{PALdsa}()},
\code{\link{nonAanalysis}()},
\code{\link{plotGenePAnumbers}()},
\code{\link{plotPADistribution}()},
\code{\link{plotPALDistribution}()},
\code{\link{plotPASignals}()}
}
\concept{Visualization functions}
