% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funclib_plot.R
\name{DSA_UpsetPlot}
\alias{DSA_UpsetPlot}
\title{DSA_UpsetPlot}
\usage{
DSA_UpsetPlot(DSA_result)
}
\arguments{
\item{DSA_result}{The output of PALdsa().}
}
\value{
A UpsetPlot.
}
\description{
Draw a Upset diagram of the result of differential tail length
  analysis.
}
\details{
Overlapping analysis of the results of differential tail length
  analysis of the four difference significance testing methods can help users
  quickly determine which results of differential tail length analysis are
  highly confident. The consensus results of the four methods are marked in
  red to represent highly confident tail length genes.
}
\examples{
library(TxDb.Mmusculus.UCSC.mm10.knownGene)
gff <- parseGenomeAnnotation(TxDb.Mmusculus.UCSC.mm10.knownGene)
data(AnnotedTails)
files = system.file("extdata", "./output/PAs/PAs.txt", package = "PolyAtailor", mustWork = TRUE)
PAs <- read.table(files,header=TRUE,sep=" ")
DSA_result <- PALdsa(PAdf=PAs,PALdf=AnnotedTails,d=24,gff=gff,mode="PD",SAoDMethod="ME",withViolinPlot=T,withUpsetPlot=F)
p <- DSA_UpsetPlot(DSA_result)
}
