% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funclib_PolyAtailor.R
\name{faBuilder}
\alias{faBuilder}
\title{faBuilder}
\usage{
faBuilder(
  fastqfile,
  mcans,
  findUmi,
  lumi,
  adapterSeq,
  anchorSeq,
  resultpath,
  samplename,
  tailAnchorLen,
  mapping,
  findTailType
)
}
\arguments{
\item{fastqfile}{The path of fastqfile.}

\item{mcans}{The maximum allowable mismatch number in the sliding window
algorithm,default=5.}

\item{findUmi}{Boolean value.Indicates whether the sequence structure
contains UMI or barcode.If it is ture, the UMI or Barcode will be extracted
separately.}

\item{lumi}{The length of umi in reads. "lumi = 0" means there is no need to
extract umi from reads.}

\item{adapterSeq}{character.If you enter a FASTQ file that does not
remove the 3 'adapter, please provide the full sequence of adapters.}

\item{anchorSeq}{character.If your sequence structure has a sequence of
anchor points identifying tails, enter this parameter.}

\item{resultpath}{The path where you want to store the result data.}

\item{samplename}{Specify a sample name for your data.}

\item{tailAnchorLen}{Specifies the minimum tail anchor point
length,default=8.}

\item{mapping}{Boolean value.The default value is F.}

\item{findTailType}{Used to specify the type of tail to extract from the
sequence. There are three values, 'A' means to search for poly(A) tails,
'T' means to search for poly(T) tails or 'both' means to search at the same
time. The default is 'both'.}

\item{minTailLen}{Specifies the minimum tail length,default=8.}
}
\value{
Generate a FASTA file for sequence alignment and save it to the
  specified directory.
}
\description{
\code{faBuilder} Tails and partial sequences were extracted from
  long reads and FASTA files were generated for alignment.
}
\details{
This function is used to extract the PolyA tail and part of the
  sequence before the tail starting site from the FASTQ file containing
  Longread, and generate the FASTA file that can be input into the sequence
  alignment software.
}
\examples{
fastqfile <- "./inst/extdata/GV_fastq/PAIso_GV1.fastq"
faBuilderRE <- faBuilder(fastqfile,
  mcans = 5, findUmi = F, resultpath = "./inst/extdata/output/", samplename =
    "GV1", tailAnchorLen = 8, mapping = F
)
head(faBuilderRE)
}
\seealso{
[tailMap()] to quantitative tails based on sequence algin.

Other Poly(A) Tail length quantification functions: 
\code{\link{tailMap}()},
\code{\link{tailScan}()}
}
\concept{Poly(A) Tail length quantification functions}
