% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funclib_PolyAtailor.R
\name{findPAs}
\alias{findPAs}
\title{findPAs}
\usage{
findPAs(chrinfo, bamfile, resultpath, count)
}
\arguments{
\item{chrinfo}{The path of chrinfo file.}

\item{bamfile}{The path of bam file.}

\item{resultpath}{The path of result file.}

\item{count}{Boolean parameter. If count=T, the PA list after the count will
be returned. There will be one more column of count in the PA list. If
count=F, all PA site information is returned without counting.}
}
\value{
Save the calculated PA site results to the result path that you specify.
}
\description{
Each chromosome was traversed and PA sites were identified by
  a-rich at the end. Note that the BAM file needs to be indexed.
}
\details{
The findPAs function takes the information from the BAM file and the
  CHRinfo file that you input, calculates the PA site for each read based on
  the A-rich in the sequence, and consolidates and saves the PA sites
  information to the result path that you specify.
}
\examples{
bamfilepath = "./data/GV_algin/PAIso-GV1.sorted.bam"
chrinfopath = "./data/GV_algin/chrinfo.txt"
resultpath = "./data/output/PAs"
test <- findPAs(chrinfo=chrinfopath,bamfile=bamfilepath,resultpath=resultpath)
}
