% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funclib_PolyAtailor.R
\name{tailMap}
\alias{tailMap}
\title{tailMap}
\usage{
tailMap(
  bamfile,
  mcans,
  minTailLen,
  findUmi,
  maxNtail,
  mapping,
  longRead,
  directBam,
  findTailType
)
}
\arguments{
\item{bamfile}{The path of bamfile.}

\item{mcans}{The maximum allowable mismatch number in the sliding window
algorithm,default=5.}

\item{minTailLen}{Specifies the minimum tail length,default=8.}

\item{findUmi}{Boolean value.Indicates whether the sequence structure
contains UMI or barcode.If it is ture, the UMI or Barcode will be extracted
separately.}

\item{maxNtail}{Specifies the maximum number of tails that should be found in
a sequence,default=2.}

\item{mapping}{Boolean value. The default value is F.}

\item{longRead}{Boolean value. If your sequence type is longreads this
parameter is T, otherwise it is F, and the default is T.}

\item{directBam}{Logical parameter. If you want to skip the tail_scan step
and directly use tail_map to analyze the compared bam file, set this
parameter to true, and the default value is true.}

\item{findTailType}{Used to specify the type of tail to extract from the
sequence. There are three values, 'A' means to search for poly(A) tails,
'T' means to search for poly(T) tails or 'both' means to search at the same
time. The default is 'both'.}
}
\value{
A dataframe with all the information of the tail.Include at least
  read_num,tail,PAL,chr,strand,tailType,read_type and sample.
}
\description{
\code{tailMap}Extract necessary comment information from BAM
  files and remove IP (internal priming).
}
\details{
If your sequence is longreads, this function is
  used to parse BAM files, remove the IP (internal priming) from the initial
  tail based on the information in the cigar field in the BAM files, and
  finally return the dataframe with all the information of the tail.However,
  if your sequence type is shortreads, this function is used to extract coord
  annotation information from BAM files.
}
\examples{
bamfile <- "./inst/extdata/GV_algin/PAIso-GV1.sorted.bam"
GV1tailMapre <- tailMap(bamfile, mcans = 5, minTailLen = 8, findUmi = F, longRead = T)
}
\seealso{
[tailScan()] to quantitative tails without sequence algin.

Other Poly(A) Tail length quantification functions: 
\code{\link{faBuilder}()},
\code{\link{tailScan}()}
}
\concept{Poly(A) Tail length quantification functions}
