% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funclib_PolyAtailor.R
\name{tailScan}
\alias{tailScan}
\title{Use a variety of methods to help you quantify the tails in a sequence.}
\usage{
tailScan(
  fastq,
  mcans,
  findUmi,
  lumi,
  adapterSeq,
  anchorSeq,
  resultpath,
  samplename,
  tailAnchorLen,
  minTailLen,
  realTailLen,
  maxNtail,
  mapping,
  mapinfo,
  findTailType
)
}
\arguments{
\item{fastq}{The path of fastqfile.}

\item{mcans}{The maximum allowable mismatch number in the sliding window
algorithm,default=5.}

\item{findUmi}{Boolean value.Indicates whether the sequence structure
contains UMI or barcode.If it is ture, the UMI or Barcode will be extracted
separately.}

\item{lumi}{The length of umi in reads. "lumi = 0" means there is no need to
extract umi from reads.}

\item{adapterSeq}{character.If you enter a FASTQ file that does not
remove the 3 'adapter, please provide the full sequence of adapters.}

\item{anchorSeq}{character.If your sequence structure has a sequence of
anchor points identifying tails, enter this parameter.}

\item{resultpath}{The path where you want to store the result data.}

\item{samplename}{Specify a sample name for your data.}

\item{tailAnchorLen}{Specifies the minimum tail anchor point
length,default=8.}

\item{minTailLen}{Specifies the minimum tail length,default=8.}

\item{realTailLen}{Specifies what you think is the true tail
length,default=30.}

\item{maxNtail}{Specifies the maximum number of tails that should be found in
a sequence,default=2.}

\item{mapping}{Boolean value.The default value is F.}

\item{findTailType}{Used to specify the type of tail to extract from the
sequence. There are three values, 'A' means to search for poly(A) tails,
'T' means to search for poly(T) tails or 'both' means to search at the same
time. The default is 'both'.}
}
\value{
Save the quantitative tail results table of various algorithms to the
  path you specify.Meanwhile, return the tail dataframe
}
\description{
\code{tailScan} Returns a table containing at least read_num, tail length,
and tail sequence.
}
\details{
This function quantifies the possible tails of all sequences in a
  FASTQ file with a non-aligned manner.You need to specify the parameters
  according to the structure of your sequence.We will save the found tail
  data and the sequence data that did not find the tail to the path you
  specified
}
\examples{
fastqfile <- "./inst/extdata/GV_fastq/PAIso_GV1.fastq"
GV1tailDF<-tailScan(fastqfile,mcans=5,findUmi = F,resultpath =
"./inst/extdata/output/",samplename =
"GV1",tailAnchorLen=8,minTailLen=8,realTailLen=20,maxNtail=2,mapping=F)
head(GV1tailDF)
}
\seealso{
[tailMap()] to quantitative tails based on sequence algin.

Other Poly(A) Tail length quantification functions: 
\code{\link{faBuilder}()},
\code{\link{tailMap}()}
}
\concept{Poly(A) Tail length quantification functions}
